# shared/openapi-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FC27 Random Team & League Manager API",
    "description": "Backend for the FC27 random team draw and friend-league tracker",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/api/teams/{id}": {
      "get": {
        "tags": [
          "Teams"
        ],
        "summary": "Get a single team by id — public",
        "operationId": "getTeam",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Teams"
        ],
        "summary": "Replace every field of a team — ADMIN only",
        "operationId": "replaceTeam",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Teams"
        ],
        "summary": "Deactivate a team (soft delete — preserves match history) — ADMIN only",
        "operationId": "deactivateTeam",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Teams"
        ],
        "summary": "Update only the given fields of a team (null = unchanged) — ADMIN only",
        "operationId": "patchTeam",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/teams": {
      "get": {
        "tags": [
          "Teams"
        ],
        "summary": "List teams, optionally filtered by star level(s) and/or league type — public, used by the wheel including for guests. Only active teams unless includeInactive=true (admin management view).",
        "operationId": "listTeams",
        "parameters": [
          {
            "name": "stars",
            "in": "query",
            "description": "Comma-separated star levels, e.g. 2,3,4",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          {
            "name": "leagueTypeId",
            "in": "query",
            "description": "Id from GET /api/league-types",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "includeInactive",
            "in": "query",
            "description": "Include deactivated teams too — for the admin management view",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Teams"
        ],
        "summary": "Create a team — ADMIN only",
        "operationId": "createTeam",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "TeamCreateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "code": {
            "type": "string",
            "minLength": 1,
            "pattern": "[A-Za-z0-9]{1,5}"
          },
          "crestUrl": {
            "type": "string"
          },
          "colorHex": {
            "type": "string",
            "pattern": "#[0-9A-Fa-f]{6}"
          },
          "starLevel": {
            "type": "integer",
            "format": "int32",
            "maximum": 5,
            "minimum": 2
          },
          "leagueTypeId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "leagueTypeId"
        ]
      },
      "TeamResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "crestUrl": {
            "type": "string"
          },
          "colorHex": {
            "type": "string"
          },
          "starLevel": {
            "type": "integer",
            "format": "int32"
          },
          "leagueTypeId": {
            "type": "integer",
            "format": "int64"
          },
          "leagueTypeName": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "TeamPatchRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "code": {
            "type": "string",
            "pattern": "[A-Za-z0-9]{1,5}"
          },
          "crestUrl": {
            "type": "string"
          },
          "colorHex": {
            "type": "string",
            "pattern": "#[0-9A-Fa-f]{6}"
          },
          "starLevel": {
            "type": "integer",
            "format": "int32",
            "maximum": 5,
            "minimum": 2
          },
          "leagueTypeId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}